Fix audio loop offset formula.

+++ Assets/Scripts/BeatManager.cs
@@ -17,7 +17,7 @@ public class BeatManager : MonoBehaviour {
 
         foreach (var beatLine in beatLines) {
             beatLine.beatRate = bpm / 60f;
-            beatLine.Run(offset / bpm / beatLine.beatRate, audioSource);
+            beatLine.Run(offset / (float) bpm / beatLine.beatRate, audioSource);
         }
 
         audioSource.Play();
